+++ Themis/Models/Exam.swift
@@ -13,10 +13,10 @@ struct Exam: Codable {
     let startDate: String?
     let endDate: String?
     let publishResultsDate: String?
-    var exerciseGroups: [ExerciseGroup]? //I don't know why this is an Array and wtf?
+    var exerciseGroups: [ExerciseGroup]? // I don't know why this is an Array and wtf?
     
     var exercises: [Exercise] {
-        let exercises = exerciseGroups?.reduce([]) { return $0 + ($1.exercises ?? []) } as? [Exercise]
+        let exercises = exerciseGroups?.reduce([]) { $0 + ($1.exercises ?? []) } as? [Exercise]
         guard let exercises else { return [] }
         return exercises.map { exercise in
             var exerciseWithDate = exercise
